Fine implementazione, aggiunta degli shoertcode su fase e attivita, inserimento di controlli su $entry_gforms per far si che non dia nessun warning anche se non ci sono entries e creazione delle classi legate alla visualizzazione dell'organigramma dei dipendenti e dei processi

+++ web/wp-content/plugins/mappaturePlugin/shortcodes/ShortCodesProcesso.php
@@ -1,45 +1,54 @@
 <?php
+
 namespace MappaturePlugin;
 
 use GFAPI;
-class ShortCodesProcesso{
+
+class ShortCodesProcesso
+{
 
     public static function create_processo()
     {
-        $lastEntry = GFAPI::get_entries(1)[0];
+        $lastEntry = GFAPI::get_entries(1);
         $processo = new Processo();
-        foreach ($lastEntry as $key => $value) {
-            $pattern = "[^9.]";
-            if (preg_match($pattern, $key) && $value) {
+        if (!empty($lastEntry)) {
+            $lastEntry = GFAPI::get_entries(1)[0];
+            foreach ($lastEntry as $key => $value) {
+                $pattern = "[^9.]";
+                if (preg_match($pattern, $key) && $value) {
 
-                $processo->setNomeProcesso($value);
-                $processo->setIdForm($lastEntry['form_id']);
-                $processo->setAreaProcesso($lastEntry[2]);
-                $processo->setServizioProcesso($lastEntry[3]);
-                $processo->setUfficioProcesso($lastEntry[4]);
-                $id_owner = GetterIdUsers::getProcessOwnerId($processo->getAreaProcesso());
-                if ($id_owner == NULL || $id_owner == '') {
-                    $id_owner = GetterIdUsers::getProcedureOwnerId($processo->getAreaProcesso(), $processo->getServizioProcesso(), $processo->getUfficioProcesso());
+                    $processo->setNomeProcesso($value);
+                    $processo->setIdForm($lastEntry['form_id']);
+                    $processo->setAreaProcesso($lastEntry[2]);
+                    $processo->setServizioProcesso($lastEntry[3]);
+                    $processo->setUfficioProcesso($lastEntry[4]);
+                    $id_owner = GetterIdUsers::getProcessOwnerId($processo->getAreaProcesso());
+                    if ($id_owner == NULL || $id_owner == '') {
+                        $id_owner = GetterIdUsers::getProcedureOwnerId($processo->getAreaProcesso(), $processo->getServizioProcesso(), $processo->getUfficioProcesso());
+                    }
+                    $processo->setIdUser($id_owner);
+                    $processo->setRuoloUser('project-manager');
+                    $processo->creaProcesso();
                 }
-                $processo->setIdUser($id_owner);
-                $processo->setRuoloUser('project-manager');
-                $processo->creaProcesso();
             }
+            $processo->findProject();
+            $processo->assignUsers($id_owner);
         }
-        $processo->findProject();
-        $processo->assignUsers($id_owner);
-
 
 
     }
+
     public static function delete_processo()
     {
         $entry_gforms = GFAPI::get_entries(1);
-        $id_current_form = $entry_gforms[0]['id'];
-        $process = new Processo();
-        $process->setNomeProcesso($entry_gforms[0][1]);
-        $process->cancellaProcesso();
-        $result = GFAPI::delete_entry($id_current_form);
+        if (!empty($entry_gforms)) {
+            $id_current_form = $entry_gforms[0]['id'];
+            $process = new Processo();
+            $process->setNomeProcesso($entry_gforms[0][1]);
+            $process->cancellaProcesso();
+            $result = GFAPI::delete_entry($id_current_form);
+        }
+
 
     }
